Fix ghost pages when deleting from XML dialog
When we delete the first (viewport) page, we must recalculate document
bounds in order to correctly decide whether the origin of the document
is contained in any of the other preexisting pages. Otherwise, the
NamedView will believe that since the origin of the document does not
lie inside any SPPage, it needs to display the "viewport" canvas item.
This canvas item was manifesting itself as a ghost page of the fixed
issue.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4885