Fix dialog category label alignment
Setting the dialog category label through the `set_label` method
inserted the label into the `PopoverMenuItem` with the `halign`
property set to `Gtk::ALIGN_FILL`. This happened because the
`PopoverMenuItem` requires the `text` parameter during instantiation
for it to add the `Label` as one of its children (which was not
specified in case of dialog category items).

This commit introduces the `set_label` method in the `PopoverMenuItem`
class which inserts the missing `Gtk::Label` child in the dialog
category `PopoverMenuItem` entries.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4925