Move "items under cursor" out of ContextMenu
The context menu was calculation which items were under the cursor based on the
last time the mouse was on the canvas. This makes sense for the right click
on the canvas. But no sense at all for the layers and objects dialog.

We move the calculation for "items under cursor" out of the ContextMenu and
let the item-tool tell us about the items under the cursor.

Then we let the layers dialog talk about the items under its cursor.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/5047