Fix creation of new mesh gradient from shared mesh.
When editing a mesh shared between more than one object, a new mesh
gradient is created. The XML tree is duplicated but the SPObject was
not being created properly.

In additon, the creation of the SPMeshNodeArray was partially dependent
on code in the class GrDrag which is a GUI element. This created an
unneeded GUI dependency.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/766