Use the start pos when hovering over the green anchor
The position of the green anchor can be weird sometimes and there
is certainly other bugs in this code. But this minor fix will
make the way the finalisation work the same way as the motion event
curve generation. Even if the anchor's position is broken.

Fix https://gitlab.com/inkscape/inkscape/-/issues/4322