Remove temporary transform when building a star
When making a star using randomness, the center position would
be used as a seed to generate the variations in the resulting shape.

But, the shape tool started the new xml with an inverse of it's parent's
transformation so the points on dragging would match and then once finished
it removed this temporary transform.

This changes the center position in documents that aren't in px or where the
shape is being drawn in a layer that has a scale transform on it.

To fix all this odd behaviour, we remove the temporary transform and instead
apply the point to item affine as we drag. Moving the stroke style adjustment
from final to construction too.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3924