Fix the confusing UI for batch export
The placement of the prefix text box and the use of the Directory
selector drop down did not create a good UX.

This is a partial reversion, using a text box with a button to
select the directory.

But this doesn't use the Save dialog to select a filename and
keeps the prefix seperated.