Fix segmentation faults with FontSelector.
Use auto_connection to delete idle connection.

An idle function is used to speed up construction of the FontSelector
widget by delaying the formatting done by the cell renderer. The
'C'-style idle function was taking "this" as a parameter. If the idle
function was still active when the FontSelector was deleted, a
segmentaiton fault resulted.

This replaces the 'C' style function with a 'C++' function, and
uses auto_connection to ensure that the function is not called
after the FontSelector is deleted.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4916