Remove GC from SPDocument.
GC was only being used to verify that the reference count was zero
during SPDocument destruction. There is no reason for SPDocument
to be garbage collected.

In order to keep the GC dependency, a std::default_delete was added
to the code. This resulted in undefined behavior and crashes. This
commit removes that function.

Possible fix to https://gitlab.com/inkscape/inkscape/-/issues/4839