Windows MSI installer: Fix upgrading from old MSI
When
1. Inkscape was installed from .MSI from Inkscape 1.3.x
2. A newer Inkscape 1.4.x was built with CMake 3.29.{0...4} and then installed from .MSI

then the user had two half-broken installations instead of one new Inkscape version.

This is due to a CMake regression
https://gitlab.kitware.com/cmake/cmake/-/issues/26029

The underlying change in CMake had originally intended to solve
https://gitlab.com/inkscape/inkscape/-/issues/1617
https://gitlab.com/inkscape/inkscape/-/issues/1618
but was effectively reverted in CMake 3.29.5 because it created these half-broken installs.

Now we effectively re-open the two above bugs to avoid the new half-broken installation.