Fix alignment and surface size of anchor/handle graphics
Related to https://gitlab.com/inkscape/inkscape/-/merge_requests/5624

Previous work to align handle graphics to pixel boundries created a situation
where the handle sizes as rendered and the surface size provisioned for the
rendering were always 3px off regardless of size. This means handles would
always be misaligned towards the top-left, worse at smaller handle sizes as
the 3px of nothing would take up more and more of the total size allotted.

Trying to pixel align by setting the _bounds box was a mistake, since that
attribute is for setting which parts of the screen need to be redrawn, it
is not a value to be used for setting the position of graphics. So a new
variable _pos is created which makes it clearer what is happening.

This allows pos and bounds to be rounded differently and at the right times
to avoid a lot of the fudge factors that also caused misalignments.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4758