Add extra path info when two Recent File entries have the same file name.
On Linux, files with indentical names could be distinguished though tooltips
that show the full path to a file. This never worked on MacOS. It might be
possible to restore this behavior on Linux but the code would be quite
complicated and not-localized. It might even be possible to add it to MacOS
but it would require low-level code.

This code works by comparing paths for files with indentical file names
and prepending the first difference to the file name.

The previous code also used the "Short File Name which removed the
added date information to the end of the file name for files created
through crashes.

Note, only when two files are found with identical file names is the
prepending code activated.