Add a line length param when in node editing mode
Add a L: [Length] spin button to the node toolbar which when
changed will modify the distance between two points to that length.

This does it in the most non-disruptive way by scaling.

A scale is applied against a center point, this is either the midpoint
of the nodes if both nodes were selected at the same time, or the
first node position if it was selected one at a time.