Fixes problem when users customise the units xml
The units xml allows us to add units dyanamically, this works because
inkscape stores all it's number as "user units" which don't record the
real-world unit but are sized by the viewBox. There is one exception
which is the document's width and height. This MUST always be an SVG
unit as defined in our SVGLength code.

Choosing a custom unit previously would reset the unit to px and cause
breakage and undefined behavour.

This fix includes:

 * Adding ft and meters as custom, non-SVGLength units
 * Adding that suppoort the pages regex.
 * Checking SVGLength when setting the unit menu list.