Don't set gradients when setting stroke styles
The stroke style shouldn't be consumed by GrDrag::styleSet this
causes a mass of problems because of how the style setters are
structured and will need to be refacted in the future.

Also reverted is a change to the way unselected gradient stops.
We don't want to select the first stop when nothing is selected.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/5165