Fix regression in straight line on node deletion
The refactoring to the node tool's delete function caused
a regression where it was impossible to delete nodes and
have the resulting line be straight.

This adds in the missing functionality as well as fixing the
ctrl modifier setting which was stolen by the new curve mode
we use shift for this new mode instead.

This is a modified backport of MR6634

Fixes https://gitlab.com/inkscape/inkscape/-/issues/5183