Rewrite of the debian packaging
This is a complete rewrite of the debian generation, to be used in the
ppa (and could also be picked up by maintainers). It splits the package
into several packages: lib2geom$VERSION and lib2geom-dev are now separate
and we also ship separately the translations, themes, extensions, and
extension manager. This uses the dh-sequence-cpack to leverage the
"COMPONENTS" of cmake.