Fix marker rendering order by controlling their z_order
There are two main problems, fistly we were prepending markers to
a list of renderable children which caused an reverse order to that
intended in the SVG specification and counter to the PDF export of
the same files.

The second problem is what happened on update. It would prepend
or append the item blindly to the list, causing edits to change
the rendering order of markers. This is fixed by moving rendered
markers into the corect z-order, even if their drawing items
haven't been created in the last update.

Fixes: https://gitlab.com/inkscape/inkscape/-/issues/5209